java path fixed, 0001_init removed, py27 updated

+++ third_parties/python/Python27/Lib/test/test_macos.py
@@ -8,9 +8,11 @@ MacOS = test_support.import_module('MacOS')
 TESTFN2 = test_support.TESTFN + '2'
 
 class TestMacOS(unittest.TestCase):
-    @unittest.skipUnless(os.path.exists('/Developer/Tools/SetFile'),
-                         '/Developer/Tools/SetFile does not exist')
+
     def testGetCreatorAndType(self):
+        if not os.path.exists('/Developer/Tools/SetFile'):
+            return
+
         try:
             fp = open(test_support.TESTFN, 'w')
             fp.write('\n')
@@ -21,15 +23,16 @@ class TestMacOS(unittest.TestCase):
                         test_support.TESTFN])
 
             cr, tp = MacOS.GetCreatorAndType(test_support.TESTFN)
-            self.assertEqual(tp, 'ABCD')
-            self.assertEqual(cr, 'EFGH')
+            self.assertEquals(tp, 'ABCD')
+            self.assertEquals(cr, 'EFGH')
 
         finally:
             os.unlink(test_support.TESTFN)
 
-    @unittest.skipUnless(os.path.exists('/Developer/Tools/GetFileInfo'),
-                         '/Developer/Tools/GetFileInfo does not exist')
     def testSetCreatorAndType(self):
+        if not os.path.exists('/Developer/Tools/GetFileInfo'):
+            return
+
         try:
             fp = open(test_support.TESTFN, 'w')
             fp.write('\n')
@@ -39,8 +42,8 @@ class TestMacOS(unittest.TestCase):
                     'ABCD', 'EFGH')
 
             cr, tp = MacOS.GetCreatorAndType(test_support.TESTFN)
-            self.assertEqual(cr, 'ABCD')
-            self.assertEqual(tp, 'EFGH')
+            self.assertEquals(cr, 'ABCD')
+            self.assertEquals(tp, 'EFGH')
 
             data = subprocess.Popen(["/Developer/Tools/GetFileInfo", test_support.TESTFN],
                     stdout=subprocess.PIPE).communicate()[0]
@@ -53,8 +56,8 @@ class TestMacOS(unittest.TestCase):
                 if ln.startswith('creator:'):
                     cr = ln.split()[-1][1:-1]
 
-            self.assertEqual(cr, 'ABCD')
-            self.assertEqual(tp, 'EFGH')
+            self.assertEquals(cr, 'ABCD')
+            self.assertEquals(tp, 'EFGH')
 
         finally:
             os.unlink(test_support.TESTFN)
@@ -74,14 +77,14 @@ class TestMacOS(unittest.TestCase):
             fp = open(test_support.TESTFN, 'r')
             data = fp.read()
             fp.close()
-            self.assertEqual(data, 'hello world\n')
+            self.assertEquals(data, 'hello world\n')
 
             rfp = MacOS.openrf(test_support.TESTFN, '*rb')
             data = rfp.read(100)
             data2 = rfp.read(100)
             rfp.close()
-            self.assertEqual(data, 'goodbye world\n')
-            self.assertEqual(data2, '')
+            self.assertEquals(data, 'goodbye world\n')
+            self.assertEquals(data2, '')
 
 
         finally:
